# spread >10x faster