--- a/tests/benchmark_masks.nim
+++ b/tests/benchmark_masks.nim
@@ -33,5 +33,31 @@ timeIt "ceil":
 
 reset()
 
-timeIt "spread":
-  mask.spread(10)
+block spread_1:
+  var p: Path
+  p.rect(500, 500, 500, 500)
+
+  timeIt "spread_1":
+    mask.fill(0)
+    mask.fillPath(p)
+    mask.spread(10)
+
+block spread_2:
+  var p: Path
+  p.rect(500, 500, 1000, 1000)
+
+  timeIt "spread_2":
+    mask.fill(0)
+    mask.fillPath(p)
+    mask.spread(10)
+
+block spread_3:
+  timeIt "spread_3":
+    mask.fill(255)
+    mask.spread(10)
+
+block spread_4:
+  timeIt "spread_4":
+    mask.fill(0)
+    mask.setValueUnsafe(1000, 1000, 255)
+    mask.spread(10)