before:
```
name ............................... min time      avg time    std dv   runs
spread_1 ......................... 755.274 ms    755.672 ms    ±0.259     x7
spread_2 ......................... 603.431 ms    603.706 ms    ±0.253     x9
spread_3 ........................... 6.811 ms      6.889 ms    ±0.126   x723
spread_4 ......................... 813.653 ms    814.233 ms    ±0.557     x7
```

after:
```
name ............................... min time      avg time    std dv   runs
spread_1 .......................... 49.026 ms     50.559 ms    ±0.793    x99
spread_2 .......................... 40.273 ms     41.585 ms    ±0.521   x121
spread_3 ........................... 8.260 ms      8.308 ms    ±0.068   x601
spread_4 .......................... 51.267 ms     53.149 ms    ±0.899    x94
```